Builder refactor

+++ src/components/manager.ts
@@ -22,8 +22,6 @@ import {FinderUtils} from './managerlib/finderutils'
 import {PathUtils} from './managerlib/pathutils'
 import {IntellisenseWatcher} from './managerlib/intellisensewatcher'
 
-import {Mutex} from '../lib/await-semaphore'
-
 /**
  * The content cache for each LaTeX file `filepath`.
  */
@@ -97,7 +95,6 @@ export class Manager implements IManager {
     private readonly finderUtils: FinderUtils
     private readonly pathUtils: PathUtils
     private readonly filesWatched = new Set<string>()
-    private readonly fileWatcherMutex = new Mutex()
     private readonly watcherOptions: chokidar.WatchOptions
     private readonly rsweaveExt: string[] = ['.rnw', '.Rnw', '.rtex', '.Rtex', '.snw', '.Snw']
     private readonly jlweaveExt: string[] = ['.jnw', '.jtexw']
@@ -600,7 +597,7 @@ export class Manager implements IManager {
             // The file is considered for the first time.
             // We must add the file to watcher to make sure we avoid infinite loops
             // in case of circular inclusion
-            await this.addToFileWatcher(file)
+            this.addToFileWatcher(file)
         }
         let content = this.getDirtyContent(file)
         if (!content) {
@@ -771,7 +768,7 @@ export class Manager implements IManager {
                 await this.parseFileAndSubs(inputFile, texFile)
             } else if (!this.filesWatched.has(inputFile)) {
                 // Watch non-tex files.
-                await this.addToFileWatcher(inputFile)
+                this.addToFileWatcher(inputFile)
             }
         }
 
@@ -810,28 +807,24 @@ export class Manager implements IManager {
     private async initiateFileWatcher() {
         await this.resetFileWatcher()
         if (this.rootFile !== undefined) {
-            await this.addToFileWatcher(this.rootFile)
+            this.addToFileWatcher(this.rootFile)
         }
     }
 
-    private async addToFileWatcher(file: string) {
-        const release = await this.fileWatcherMutex.acquire()
-        try {
-            this.fileWatcher.add(file)
-            this.filesWatched.add(file)
-        } finally {
-            release()
+    private addToFileWatcher(file: string) {
+        if (this.filesWatched.has(file)) {
+            return
         }
+        this.filesWatched.add(file)
+        this.fileWatcher.add(file)
     }
 
-    private async deleteFromFileWatcher(file: string) {
-        const release = await this.fileWatcherMutex.acquire()
-        try {
-            this.fileWatcher.unwatch(file)
-            this.filesWatched.delete(file)
-        } finally {
-            release()
+    private deleteFromFileWatcher(file: string) {
+        if (!this.filesWatched.has(file)) {
+            return
         }
+        this.filesWatched.delete(file)
+        this.fileWatcher.unwatch(file)
     }
 
     private createFileWatcher() {
@@ -849,15 +842,10 @@ export class Manager implements IManager {
     }
 
     private async resetFileWatcher() {
-        const release = await this.fileWatcherMutex.acquire()
-        try {
-            this.extension.logger.addLogMessage('Reset file watcher.')
-            await this.fileWatcher.close()
-            this.registerListeners(this.fileWatcher)
-            this.filesWatched.clear()
-        } finally {
-            release()
-        }
+        this.extension.logger.addLogMessage('Reset file watcher.')
+        await this.fileWatcher.close()
+        this.registerListeners(this.fileWatcher)
+        this.filesWatched.clear()
         // We also clean the completions from the old project
         this.extension.completer.input.reset()
         this.extension.duplicateLabels.reset()
@@ -884,9 +872,9 @@ export class Manager implements IManager {
         await this.buildOnFileChanged(file)
     }
 
-    private async onWatchedFileDeleted(file: string) {
+    private onWatchedFileDeleted(file: string) {
         this.extension.logger.addLogMessage(`File watcher - file deleted: ${file}`)
-        await this.deleteFromFileWatcher(file)
+        this.deleteFromFileWatcher(file)
         delete this.cachedContent[file]
         if (file === this.rootFile) {
             this.extension.logger.addLogMessage(`Root file deleted: ${file}`)
